Time: 8 ms (83.99%), Space: 6 MB (31.16%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool isPalindrome(int x) {
+    bool isPalindrome(long int x) {
       long int rev=0, temp = x, rem=0;
       if(x<0 || (x!=0 && x%10==0) ) return false;
       while(temp>0)
